+++ data-access/src/main/java/com/healthplus/dataaccess/domain/DietPlan.java
@@ -15,7 +15,7 @@ public class DietPlan implements Serializable {
     
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)
-    private int id;
+    private Long id;
 
     @NotNull
     private Prescription prescription;
@@ -24,13 +24,13 @@ public class DietPlan implements Serializable {
     private String food;
     
     @Positive
-    private int duration;
+    private Integer duration;
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -50,11 +50,11 @@ public class DietPlan implements Serializable {
 		this.food = food;
 	}
 
-	public int getDuration() {
+	public Integer getDuration() {
 		return duration;
 	}
 
-	public void setDuration(int duration) {
+	public void setDuration(Integer duration) {
 		this.duration = duration;
 	}
 }
