+++ data-access/src/main/java/com/healthplus/dataaccess/domain/OccupiedBed.java
@@ -16,7 +16,7 @@ public class OccupiedBed implements Serializable {
 
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)
-    private int id;
+    private Long id;
 
     @NotNull
     private Bed bed;
@@ -31,13 +31,13 @@ public class OccupiedBed implements Serializable {
 
     @NotNull
     @Positive
-    private int rate;
+    private Integer rate;
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -73,11 +73,11 @@ public class OccupiedBed implements Serializable {
 		this.endTime = endTime;
 	}
 
-	public int getRate() {
+	public Integer getRate() {
 		return rate;
 	}
 
-	public void setRate(int rate) {
+	public void setRate(Integer rate) {
 		this.rate = rate;
 	}
 }
