+++ data-access/src/main/java/com/healthplus/dataaccess/domain/Scheme.java
@@ -14,7 +14,7 @@ public class Scheme implements Serializable {
 
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)
-    private int id;
+    private Long id;
 
     @NotNull
     private String name;
@@ -22,13 +22,13 @@ public class Scheme implements Serializable {
     private String description;
 
     @NotNull
-    private int amountRedeemable;
+    private Integer amountRedeemable;
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -48,11 +48,11 @@ public class Scheme implements Serializable {
 		this.description = description;
 	}
 
-	public int getAmountRedeemable() {
+	public Integer getAmountRedeemable() {
 		return amountRedeemable;
 	}
 
-	public void setAmountRedeemable(int amountRedeemable) {
+	public void setAmountRedeemable(Integer amountRedeemable) {
 		this.amountRedeemable = amountRedeemable;
 	}
 }
